remove missing_trie_keys, and digests_without_missing_descendants cache
