+++ execution_engine/src/storage/global_state/in_memory.rs
@@ -21,8 +21,8 @@ use crate::{
         trie_store::{
             in_memory::InMemoryTrieStore,
             operations::{
-                self, keys_with_prefix, missing_children, missing_trie_keys, put_trie, read,
-                read_with_proof, ReadResult, WriteResult,
+                self, keys_with_prefix, missing_children, put_trie, read, read_with_proof,
+                ReadResult, WriteResult,
             },
         },
     },
@@ -263,30 +263,6 @@ impl StateProvider for InMemoryGlobalState {
         >(correlation_id, &mut txn, &self.trie_store, trie)?;
         txn.commit()?;
         Ok(trie_hash)
-    }
-
-    /// Finds all of the keys of missing descendant `Trie<Key,StoredValue>` values.
-    fn missing_trie_keys(
-        &self,
-        correlation_id: CorrelationId,
-        trie_keys: Vec<Digest>,
-    ) -> Result<Vec<Digest>, Self::Error> {
-        let txn = self.environment.create_read_txn()?;
-        let missing_descendants = missing_trie_keys::<
-            Key,
-            StoredValue,
-            InMemoryReadTransaction,
-            InMemoryTrieStore,
-            Self::Error,
-        >(
-            correlation_id,
-            &txn,
-            self.trie_store.deref(),
-            trie_keys,
-            &Default::default(),
-        )?;
-        txn.commit()?;
-        Ok(missing_descendants)
     }
 
     /// Finds all of the keys of missing directly descendant `Trie<Key,StoredValue>` values.
