+++ execution_engine/src/storage/global_state/mod.rs
@@ -114,13 +114,6 @@ pub trait StateProvider {
     /// Insert a trie node into the trie
     fn put_trie(&self, correlation_id: CorrelationId, trie: &[u8]) -> Result<Digest, Self::Error>;
 
-    /// Finds all of the missing or corrupt keys of which are descendants of `trie_key`.
-    fn missing_trie_keys(
-        &self,
-        correlation_id: CorrelationId,
-        trie_keys: Vec<Digest>,
-    ) -> Result<Vec<Digest>, Self::Error>;
-
     /// Finds all the children of `trie_raw` which aren't present in the state.
     fn missing_children(
         &self,
