+++ execution_engine/src/storage/trie_store/operations/tests/synchronize.rs
@@ -1,3 +1,5 @@
+use std::collections::HashSet;
+
 use casper_hashing::Digest;
 use casper_types::bytesrepr::{self, FromBytes, ToBytes};
 
@@ -6,7 +8,8 @@ use crate::{
     storage::{
         error,
         error::in_memory,
-        transaction_source::{Transaction, TransactionSource},
+        transaction_source::{Readable, Transaction, TransactionSource},
+        trie::{Pointer, Trie},
         trie_store::{
             operations::{
                 self,
@@ -17,6 +20,89 @@ use crate::{
         },
     },
 };
+
+/// Given a root hash, find any trie keys that are descendant from it that are referenced but not
+/// present in the database.
+// TODO: We only need to check one trie key at a time
+// this is only
+pub fn missing_trie_keys<K, V, T, S, E>(
+    _correlation_id: CorrelationId,
+    txn: &T,
+    store: &S,
+    mut trie_keys_to_visit: Vec<Digest>,
+    known_complete: &HashSet<Digest>,
+) -> Result<Vec<Digest>, E>
+where
+    K: ToBytes + FromBytes + Eq + std::fmt::Debug,
+    V: ToBytes + FromBytes + std::fmt::Debug,
+    T: Readable<Handle = S::Handle>,
+    S: TrieStore<K, V>,
+    S::Error: From<T::Error>,
+    E: From<S::Error> + From<bytesrepr::Error>,
+{
+    let mut missing_descendants = Vec::new();
+    let mut visited = HashSet::new();
+    while let Some(trie_key) = trie_keys_to_visit.pop() {
+        if !visited.insert(trie_key) {
+            continue;
+        }
+
+        if known_complete.contains(&trie_key) {
+            // Skip because we know there are no missing descendants.
+            continue;
+        }
+
+        let retrieved_trie_bytes = match store.get_raw(txn, &trie_key)? {
+            Some(bytes) => bytes,
+            None => {
+                // No entry under this trie key.
+                missing_descendants.push(trie_key);
+                continue;
+            }
+        };
+
+        // Optimization: Don't deserialize leaves as they have no descendants.
+        if let Some(&Trie::<K, V>::LEAF_TAG) = retrieved_trie_bytes.first() {
+            continue;
+        }
+
+        // Parse the trie, handling errors gracefully.
+        let retrieved_trie = match bytesrepr::deserialize_from_slice(retrieved_trie_bytes) {
+            Ok(retrieved_trie) => retrieved_trie,
+            // Couldn't parse; treat as missing and continue.
+            Err(err) => {
+                tracing::error!(?err, "unable to parse trie");
+                missing_descendants.push(trie_key);
+                continue;
+            }
+        };
+
+        match retrieved_trie {
+            // Should be unreachable due to checking the first byte as a shortcut above.
+            Trie::<K, V>::Leaf { .. } => {
+                tracing::error!(
+                    "did not expect to see a trie leaf in `missing_trie_keys` after shortcut"
+                );
+            }
+            // If we hit a pointer block, queue up all of the nodes it points to
+            Trie::Node { pointer_block } => {
+                for (_, pointer) in pointer_block.as_indexed_pointers() {
+                    match pointer {
+                        Pointer::LeafPointer(descendant_leaf_trie_key) => {
+                            trie_keys_to_visit.push(descendant_leaf_trie_key)
+                        }
+                        Pointer::NodePointer(descendant_node_trie_key) => {
+                            trie_keys_to_visit.push(descendant_node_trie_key)
+                        }
+                    }
+                }
+            }
+            // If we hit an extension block, add its pointer to the queue
+            Trie::Extension { pointer, .. } => trie_keys_to_visit.push(pointer.into_hash()),
+        }
+    }
+    Ok(missing_descendants)
+}
 
 fn copy_state<'a, K, V, R, S, E>(
     correlation_id: CorrelationId,
@@ -37,7 +123,7 @@ where
     // Make sure no missing nodes in source
     {
         let txn: R::ReadTransaction = source_environment.create_read_txn()?;
-        let missing_from_source = operations::missing_trie_keys::<_, _, _, _, E>(
+        let missing_from_source = missing_trie_keys::<_, _, _, _, E>(
             correlation_id,
             &txn,
             source_store,
@@ -61,7 +147,7 @@ where
             target_store.put_raw(&mut target_txn, &trie_key, trie_bytes_to_insert.as_ref())?;
 
             // Now that we've added in `trie_to_insert`, queue up its children
-            let new_keys = operations::missing_trie_keys::<_, _, _, _, E>(
+            let new_keys = missing_trie_keys::<_, _, _, _, E>(
                 correlation_id,
                 &target_txn,
                 target_store,
@@ -78,7 +164,7 @@ where
     // After the copying process above there should be no missing entries in the target
     {
         let target_txn: R::ReadWriteTransaction = target_environment.create_read_write_txn()?;
-        let missing_from_target = operations::missing_trie_keys::<_, _, _, _, E>(
+        let missing_from_target = missing_trie_keys::<_, _, _, _, E>(
             correlation_id,
             &target_txn,
             target_store,
