+++ utils/retrieve-state/src/lib.rs
@@ -3,11 +3,11 @@ use std::{
     convert::TryInto,
     net::{AddrParseError, SocketAddr},
     sync::Arc,
-    time::{Duration, Instant, SystemTime, UNIX_EPOCH},
+    time::{Duration, Instant},
 };
 
 use futures::stream::FuturesUnordered;
-use futures_util::{future::Either, stream::StreamExt, FutureExt};
+use futures_util::stream::StreamExt;
 pub use reqwest::Client;
 use serde::{Deserialize, Serialize};
 
@@ -32,11 +32,8 @@ use casper_node::{
     types::{Block, BlockHash, Deploy, DeployOrTransferHash, JsonBlock},
 };
 use casper_types::{bytesrepr, bytesrepr::Bytes};
-use futures_channel::{mpsc::UnboundedSender, oneshot};
-use futures_util::SinkExt;
 use tokio::{
     sync::{mpsc, Mutex},
-    task::JoinHandle,
     time::error::Elapsed,
 };
 use tracing::{error, info, trace, warn};
@@ -432,362 +429,6 @@ impl TriesAwaitingChildren {
             ),
         );
         missing_children
-    }
-
-    fn handle_result(
-        &mut self,
-        trie_key: Digest,
-        trie_raw: TrieRaw,
-        put_result: Result<Digest, engine_state::Error>,
-    ) -> Vec<Digest> {
-        match put_result {
-            Ok(written_trie) if written_trie == trie_key => self.handle_written(written_trie),
-            Ok(written_trie) => {
-                error!(%written_trie, %trie_key, "trie was written under a different key");
-                vec![]
-            }
-            Err(engine_state::Error::MissingTrieNodeChildren(missing_children)) => {
-                self.handle_missing_children(trie_key, trie_raw, missing_children)
-            }
-            Err(error) => {
-                error!(%error, "couldn't put trie to the database");
-                vec![]
-            }
-        }
-    }
-}
-
-/// Download the trie from a node to the provided lmdb path.
-pub async fn download_trie_channels(
-    client: &Client,
-    peers: &[SocketAddr],
-    engine_state: Arc<EngineState<LmdbGlobalState>>,
-    state_root_hash: Digest,
-    peer_mailbox_size: usize,
-) -> Result<(), anyhow::Error> {
-    let (base_send, mut base_recv) = futures_channel::mpsc::unbounded();
-    let mut peer_futures = Vec::new();
-    for peer in peers {
-        let base_send = base_send.clone();
-        let peer_future = async move {
-            match maybe_construct_peer(base_send.clone(), *peer, client, state_root_hash).await {
-                Ok(peer) => Some(peer),
-                Err(_err) => {
-                    // just skip those that can't be communicated with
-                    // info!("error constructing peer with address {:?} {:?}", peer,
-                    // err);
-                    None
-                }
-            }
-        };
-        peer_futures.push(peer_future);
-    }
-    let mut peer_map = futures::future::join_all(peer_futures)
-        .await
-        .into_iter()
-        .flatten()
-        .map(|peer: Peer| (peer.address, peer))
-        .collect::<HashMap<SocketAddr, Peer>>();
-
-    info!("got {}/{} valid peers", peer_map.len(), peers.len());
-
-    let mut total_tries_count = 0;
-
-    let mut process_queue = vec![state_root_hash];
-    let mut tries_awaiting_children = TriesAwaitingChildren::new();
-    let mut tries_downloaded = 0usize;
-    let mut in_flight_counters = peers
-        .iter()
-        .map(|addr| (addr, 0usize))
-        .collect::<HashMap<_, _>>();
-
-    let start_instant = Instant::now();
-    let mut time_chunk = Instant::now();
-    let mut bytes_downloaded_since_last_iteration = 0;
-
-    // Top level task keeps track of missing descendants and work distribution.
-    loop {
-        // Distribute work
-        for (address, peer) in peer_map.iter() {
-            let counter = in_flight_counters.get_mut(&address).unwrap();
-            if *counter < peer_mailbox_size {
-                if let Some(next_trie_key) = process_queue.pop() {
-                    if let Some(trie_bytes) = tries_awaiting_children.get(&next_trie_key) {
-                        let put_result = tokio::task::block_in_place(|| {
-                            engine_state
-                                .put_trie_if_all_children_present(CorrelationId::new(), trie_bytes)
-                        });
-                        process_queue.extend(tries_awaiting_children.handle_result(
-                            next_trie_key,
-                            TrieRaw::new(trie_bytes.clone()),
-                            put_result,
-                        ));
-                    } else {
-                        let mut sender = peer.sender.clone();
-                        sender.send(PeerMsg::GetTrie(next_trie_key)).await.unwrap();
-                        *counter += 1;
-                    }
-                } else {
-                    break;
-                }
-            }
-        }
-
-        // Handle messages from workers
-        if let Some(msg) = base_recv.next().await {
-            match msg {
-                Ok(PeerMsg::TrieDownloaded {
-                    trie_key,
-                    trie_bytes,
-                    peer,
-                    elapsed: _,
-                    len_in_bytes,
-                }) => {
-                    let put_result = tokio::task::block_in_place(|| {
-                        engine_state
-                            .put_trie_if_all_children_present(CorrelationId::new(), &trie_bytes)
-                    });
-
-                    // count of all tries we know about
-                    total_tries_count += match &put_result {
-                        Err(engine_state::Error::MissingTrieNodeChildren(missing_children)) => {
-                            missing_children.len()
-                        }
-                        _ => 0,
-                    };
-                    process_queue.extend(tries_awaiting_children.handle_result(
-                        trie_key,
-                        TrieRaw::new(trie_bytes),
-                        put_result,
-                    ));
-                    bytes_downloaded_since_last_iteration += len_in_bytes;
-
-                    // in-flight requests
-                    if let Some(counter) = in_flight_counters.get_mut(&peer) {
-                        *counter = counter.saturating_sub(1);
-                    }
-
-                    // count of all tries downloaded
-                    tries_downloaded += 1;
-                }
-                Ok(PeerMsg::NoBytes { trie_key, peer }) => {
-                    info!(
-                        "got no bytes for key {} from peer {} (will query for it again...)",
-                        trie_key, peer
-                    );
-                    process_queue.push(trie_key);
-                }
-                Err(PeerError::BadData {
-                    trie_key,
-                    peer_address,
-                    error,
-                }) => {
-                    info!(
-                        "got bad data at key {} from peer from peer {} error {:?}",
-                        trie_key, peer_address, error
-                    );
-                    process_queue.push(trie_key);
-                }
-                Err(PeerError::TimedOut {
-                    trie_key,
-                    peer_address,
-                    error,
-                }) => {
-                    info!(
-                        "timed out asking for trie at key {} from peer {} error {:?}",
-                        trie_key, peer_address, error
-                    );
-                    process_queue.push(trie_key);
-                }
-                Err(PeerError::FailedRpc {
-                    trie_key,
-                    peer_address,
-                    error,
-                }) => {
-                    info!(
-                        "rpc failed to get key {} from peer {} error {:?}",
-                        trie_key, peer_address, error
-                    );
-                    process_queue.push(trie_key);
-                }
-                Err(PeerError::AddrParseError(addr_parse_error)) => {
-                    panic!("failed to parse address {:?}", addr_parse_error);
-                }
-                Ok(PeerMsg::GetTrie(_)) => panic!("GetTrie should not be sent by workers"),
-            }
-        }
-
-        if time_chunk.elapsed() > Duration::from_secs(1) {
-            let chunk_elapsed = time_chunk.elapsed().as_millis() as usize;
-            let bytes_per_sec = (bytes_downloaded_since_last_iteration / chunk_elapsed) * 1000;
-            let kbytes_per_sec = bytes_per_sec / 1024;
-            let timestamp = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .unwrap()
-                .as_millis();
-            info!(
-                "({timestamp}) tries: {downloaded}/{total} - {kbytes_per_sec} kilobytes per second",
-                downloaded = tries_downloaded,
-                total = total_tries_count,
-                kbytes_per_sec = kbytes_per_sec,
-                timestamp = timestamp,
-            );
-            bytes_downloaded_since_last_iteration = 0;
-            time_chunk = Instant::now();
-        }
-
-        // Finishing state, we downloaded all the intermediate tries we found, as well as the state
-        // root we started with.
-        if tries_downloaded == total_tries_count + 1 {
-            let missing =
-                engine_state.missing_trie_keys(CorrelationId::new(), vec![state_root_hash])?;
-
-            if missing.is_empty() {
-                info!("state root has no missing descendants.");
-                info!(
-                    "tries downloaded: {downloaded}/{total} in {total_time} seconds",
-                    downloaded = tries_downloaded,
-                    total = total_tries_count,
-                    total_time = start_instant.elapsed().as_millis(),
-                );
-                break;
-            } else {
-                info!(
-                    "state root still has {} missing descendants. {:#?}",
-                    missing.len(),
-                    in_flight_counters,
-                );
-            }
-        }
-    }
-
-    for (_address, peer) in peer_map.drain() {
-        peer.shutdown.send(()).unwrap();
-        peer.join_handle.await.unwrap();
-    }
-
-    Ok(())
-}
-
-struct Peer {
-    address: SocketAddr,
-    sender: UnboundedSender<PeerMsg>,
-    join_handle: JoinHandle<()>,
-    shutdown: oneshot::Sender<()>,
-}
-
-async fn maybe_construct_peer(
-    base_send: UnboundedSender<Result<PeerMsg, PeerError>>,
-    address: SocketAddr,
-    client: &Client,
-    state_root_hash: Digest,
-) -> Result<Peer, PeerError> {
-    let url = address_to_url(address);
-    let base_send = base_send.clone();
-    let client = client.clone();
-    let read_future = get_trie(
-        &client,
-        &url,
-        GetTrieParams {
-            trie_key: state_root_hash,
-        },
-    );
-    let read_or_timeout = tokio::time::timeout(Duration::from_secs(1), read_future).await;
-    match read_or_timeout {
-        Ok(Ok(_)) => Ok(spawn_peer(base_send, address, client)),
-        Ok(Err(error)) => Err(PeerError::FailedRpc {
-            peer_address: address,
-            trie_key: state_root_hash,
-            error,
-        }),
-        Err(elapsed) => Err(PeerError::TimedOut {
-            peer_address: address,
-            trie_key: state_root_hash,
-            error: elapsed,
-        }),
-    }
-}
-
-/// Spawns a task to handle requests sent to this peer. This task will query the peer for a given
-/// Trie under a Digest. The returned Digest will be written to LMDB global state.
-fn spawn_peer(
-    base_send: UnboundedSender<Result<PeerMsg, PeerError>>,
-    address: SocketAddr,
-    client: Client,
-) -> Peer {
-    let (sender, peer_recv) = futures_channel::mpsc::unbounded::<PeerMsg>();
-    let peer_recv = peer_recv.map(Either::Left);
-
-    let (shutdown, shutdown_recv) = futures_channel::oneshot::channel::<()>();
-    let shutdown_recv = shutdown_recv.into_stream().map(|_| Either::Right(()));
-
-    let mut next_peer_msg_or_shutdown_stream = futures::stream::select(peer_recv, shutdown_recv);
-
-    let join_handle = tokio::spawn(async move {
-        let client = client.clone();
-        'peer: while let Some(Either::Left(PeerMsg::GetTrie(next_trie_key))) =
-            next_peer_msg_or_shutdown_stream.next().await
-        {
-            let start = Instant::now();
-            let mut base_send = base_send.clone();
-            let url = address_to_url(address);
-            match get_trie(
-                &client,
-                &url,
-                GetTrieParams {
-                    trie_key: next_trie_key,
-                },
-            )
-            .await
-            {
-                Ok(GetTrieResult {
-                    maybe_trie_bytes: Some(trie_bytes),
-                    ..
-                }) => {
-                    let len_in_bytes = trie_bytes.as_ref().len();
-                    base_send
-                        .send(Ok(PeerMsg::TrieDownloaded {
-                            trie_key: next_trie_key,
-                            trie_bytes,
-                            peer: address,
-                            elapsed: start.elapsed(),
-                            len_in_bytes,
-                        }))
-                        .await
-                        .expect("should send");
-                }
-                Ok(GetTrieResult {
-                    maybe_trie_bytes: None,
-                    ..
-                }) => {
-                    // No bytes for this trie? let's try another peer.
-                    base_send
-                        .send(Ok(PeerMsg::NoBytes {
-                            trie_key: next_trie_key,
-                            peer: address,
-                        }))
-                        .await
-                        .expect("should send");
-                }
-                Err(error) => {
-                    base_send
-                        .send(Err(PeerError::FailedRpc {
-                            peer_address: address,
-                            trie_key: next_trie_key,
-                            error,
-                        }))
-                        .await
-                        .expect("should send");
-                    continue 'peer;
-                }
-            };
-        }
-    });
-    Peer {
-        address,
-        sender,
-        join_handle,
-        shutdown,
     }
 }
